Switched from module_eval to define_method.
This allows the scoping to work a lot better and eliminates the need for
the @@fields hash on EasyCmp: every class that calls easy_cmp now keeps
its own @@easycmp_fields array of fields. If it weren't for the need to
have subsequent easy_cmp fields append more fields, one could just have
"for field in fields" in #easy_cmp, but this way requires storage of
past fields. Ah well!