A bug occurred to me. This test proves it.
Using class variables is a mistake because they're shared with
subclasses. I'm going to switch to instance variables of the class
objects to fix it, so /each/ class will have its own comparators.

This opens the question of whether a subclass should use a superclass's
fields at any point, though? I'll cross that bridge when I know which
bridge to cross, though. : P