Switched fields to arrays, added #merge_fields.
This makes more sense than using hashes and will be easier to backport.
As I was going to have to do some terrible things to the hases to make
them come out in the right order anyway, I decided to take the plunge
and do the same horrible things to arrays and knock out two birds with
one stone.